requirements.txt
yfinance=0.2.3
pyyaml==6.0
pymongo==4.3.3